refactor(phase4): extract config_parsing from eon_parser.rs
Move auth, database, security, TLS, logging, build, runtime, and
authorization config parsers (~1,181 lines) into
`eon_parser/config_parsing.rs`. `validate_tls_path` imports
`validate_relative_path` from sibling `storage_parsing`.
Prune all stale imports from `eon_parser.rs` that came from this
and the previous two extractions (test-only imports retained).
All 233 passing tests remain green; 19 pre-existing failures unchanged.
